% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_robust.R
\name{geocode_robust}
\alias{geocode_robust}
\title{Robust geocoding function}
\usage{
geocode_robust(data, location = "address", n = 10,
  return_full_dataframe = FALSE)
}
\arguments{
\item{data}{the data.frame containing the location column to be geocoded.
The default is "address"}

\item{location}{the name of the location column to be geocoded (character
vector of length 1)}

\item{n}{the number of times the location column should be geocoded
The default is 10. (numeric vector of length 1)}

\item{return_full_dataframe}{if TRUE, returns the original dataframe plus
the lat and lon columns. If FALSE, returns the lat and lon columns.}
}
\value{
a data.frame identical to that of the input data frame except with
lat and lon columns for the location column or just the lat and lon columns.
}
\description{
geocode_robust geocodes locations via the Google Maps API. It then
selects the nonmissing observations for the final and returned latlons.
It continues this process until all latlons are found.
}
\note{
this function will throw an error if `data` is not a data.frame

this function will throw an error if `location` is not a non-null
column of `data`

this function will throw an error if `n' is not coercible to a
non-null integer vector of length 1
}
\examples{
\dontrun{
data <- data.frame(
addr = c("The White House", "11 Wall St, New York, NY"),
stringsAsFactors = FALSE
)
data1 <- geocode_robust(data, location = "addr")
}

}
